backend/requirements.txt
asgiref==3.6.0
autopep8==2.0.2
cffi==1.15.1
cryptography==40.0.2
dj-database-url==2.1.0
Django==4.2
django-cors-headers==3.14.0
django-filter==23.1
django-rest-knox==4.2.0
djangorestframework==3.14.0
gunicorn==21.2.0
packaging==23.2
psycopg2==2.9.6
pycodestyle==2.10.0
pycparser==2.21
pytz==2023.3
sqlparse==0.4.3
typing_extensions==4.8.0
tzdata==2023.3
